requirements.txt
djangorestframework
django-cors-headers
requests
pytest
coverage